api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "WorkWithUs API — платформа взаимодействия студентов и компаний"
readme = "README.md"
requires-python = ">=3.12"
dependencies = ["fastapi>=0.115", "uvicorn[standard]>=0.34", "tortoise-orm[asyncpg]>=0.25", "python-jose[cryptography]>=3.3", "bcrypt>=4.0", "pydantic-settings>=2.7", "email-validator>=2.0", "aerich>=0.8", "loguru>=0.7", "minio>=7.2", "python-multipart>=0.0.9"]

[tool.aerich]
tortoise_orm = "src.config.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
